refactor: 관계 설정 후 사용되는 부분들 모두 수정

+++ src/options/options.service.ts
@@ -3,6 +3,7 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Option } from './entities/options.entity';
 import { CreateOptionRequestDto } from './dtos/create-options.dro';
+import { Menu } from 'src/menus/entities/menus.entity';
 import { UpdateOptionRequestDto } from './dtos/update-options.dto';
 
 @Injectable()
@@ -10,14 +11,22 @@ export class OptionsService {
   constructor(
     @InjectRepository(Option)
     private readonly optionRepository: Repository<Option>,
+    @InjectRepository(Menu)
+    private readonly menuRepository: Repository<Menu>,
   ) {}
 
   async createOption(menuId: string, createOptionRequestDto: CreateOptionRequestDto): Promise<Option> {
-    const newOption = new Option();
+    const menu = await this.menuRepository.findOne({ where: { id: menuId } });
 
-    newOption.menu_id = menuId;
-    newOption.option_name = createOptionRequestDto.optionName;
-    newOption.option_price = createOptionRequestDto.optionPrice;
+    if (!menu) {
+      throw new NotFoundException(`${menuId}에 해당하는 메뉴를 찾지 못했습니다.`);
+    }
+
+    const newOption = this.optionRepository.create({
+      menu,
+      option_name: createOptionRequestDto.optionName,
+      option_price: createOptionRequestDto.optionPrice,
+    });
 
     return await this.optionRepository.save(newOption);
   }
@@ -27,10 +36,13 @@ export class OptionsService {
     optionId: string,
     updateOptionRequestDto: UpdateOptionRequestDto,
   ): Promise<Option> {
-    const option = await this.optionRepository.findOne({ where: { id: optionId, menu_id: menuId } });
+    const option = await this.optionRepository.findOne({
+      where: { id: optionId },
+      relations: ['menu'],
+    });
 
-    if (!option) {
-      throw new NotFoundException(`${optionId}에 해당하는 옵션이 없습니다.`);
+    if (!option || option.menu.id !== menuId) {
+      throw new NotFoundException(`${menuId}에 해당하는 옵션 ${optionId}을 찾지 못했습니다. `);
     }
 
     option.option_name = updateOptionRequestDto.optionName;
@@ -40,9 +52,13 @@ export class OptionsService {
   }
 
   async deleteOption(menuId: string, optionId: string): Promise<void> {
-    const option = await this.optionRepository.findOne({ where: { id: optionId, menu_id: menuId } });
-    if (!option) {
-      throw new NotFoundException(`메뉴 ID ${menuId}에 해당하는 옵션 ID ${optionId}를 찾을 수 없습니다.`);
+    const option = await this.optionRepository.findOne({
+      where: { id: optionId },
+      relations: ['menu'],
+    });
+
+    if (!option || option.menu.id !== menuId) {
+      throw new NotFoundException(`${menuId}에 해당하는 옵션 ${optionId}을 찾지 못했습니다. `);
     }
 
     await this.optionRepository.remove(option);
